[WIP] refactor test_nspv added assert_methods moved nspvlib to framework
